+++ src/AzzyBot.Bot/Extensions/IServiceCollectionExtensions.cs
@@ -65,7 +65,7 @@ public static class IServiceCollectionExtensions
         {
             o.AddJob<AzuraRequestJob>();
             o.AddJob<AzzyBotGlobalChecksJob>(j => j.WithCronExpression("*/15 * * * *").WithName(nameof(AzzyBotGlobalChecksJob))); // Every 15 minutes
-            o.AddJob<LogfileCleaningJob>(j => j.WithCronExpression("0 0 */1 * *").WithName(nameof(LogfileCleaningJob)).WithParameter(logDays)); // Every day
+            o.AddJob<LogfileCleaningJob>(j => j.WithCronExpression("0 0 */1 * *").WithName(nameof(LogfileCleaningJob)).WithParameter(logDays)).RunAtStartup(); // Every day
         });
         services.AddSingleton<CronJobManager>();